Refactor of forms, consolidated components and created new finer grained view components.
