+++ app/components/modules/main/Header/ActionsList/AuthActions.tsx
@@ -1,0 +1,16 @@
+export interface AuthActionsProps {
+  logout: () => void;
+}
+
+const AuthActions: React.FC<AuthActionsProps> = ({ logout }) => {
+  return (
+    <li className="mr-3">
+      <button
+        onClick={logout}
+        className="inline-block text-gray-600 no-underline hover:text-gray-200 hover:text-underline py-2 px-4">
+        Log out
+      </button>
+    </li>
+  );
+};
+export default AuthActions;
