+++ app/components/modules/main/Header/ActionsList/DefaultActions.tsx
@@ -1,0 +1,12 @@
+const DefaultActions: React.FC = () => {
+  return (
+    <li className="mr-3">
+      <a
+        className="inline-block text-gray-600 no-underline hover:text-gray-200 hover:text-underline py-2 px-4"
+        href="/">
+        Home
+      </a>
+    </li>
+  );
+};
+export default DefaultActions;
