+++ app/components/modules/main/Header/ActionsList/UnauthActions.tsx
@@ -1,0 +1,22 @@
+const UnauthActions: React.FC = () => {
+  return (
+    <>
+      <li className="mr-3">
+        <a
+          className="inline-block text-gray-600 no-underline hover:text-gray-200 hover:text-underline py-2 px-4"
+          href="/login">
+          Log in
+        </a>
+      </li>
+
+      <li className="mr-3">
+        <a
+          className="inline-block text-gray-600 no-underline hover:text-gray-200 hover:text-underline py-2 px-4"
+          href="/signup">
+          Sign up
+        </a>
+      </li>
+    </>
+  );
+};
+export default UnauthActions;
