+++ app/components/modules/registration/RegistrationPanel.tsx
@@ -1,21 +1,22 @@
 import Login from '@components/modules/registration/Login';
 import SignUp from '@components/modules/registration/SignUp';
 import Forgot from '@components/modules/registration/Forgot';
+import Logo from '@components/common/Logo';
 import { useState, useEffect, ReactElement } from 'react';
+
+const headingLookup: Map<string, string> = new Map([
+  ['/login', 'Log in below to get started'],
+  ['/signup', 'Create an account below to get started'],
+  ['/forgot', 'Enter your email to reset your password']
+]);
+const componentLookup: Map<string, ReactElement> = new Map([
+  ['/login', <Login />],
+  ['/signup', <SignUp />],
+  ['/forgot', <Forgot />]
+]);
 
 const RegistrationPanel: React.FC = () => {
   const [route, setRoute] = useState('');
-  const headingLookup: Map<string, string> = new Map([
-    ['/login', 'Log in below to get started'],
-    ['/signup', 'Create an account below to get started'],
-    ['/forgot', 'Enter your email to reset your password']
-  ]);
-  const componentLookup: Map<string, ReactElement> = new Map([
-    ['/login', <Login />],
-    ['/signup', <SignUp />],
-    ['/forgot', <Forgot />]
-  ]);
-
   useEffect(() => {
     const pathName = window && window.location.pathname;
     setRoute(pathName);
@@ -25,10 +26,7 @@ const RegistrationPanel: React.FC = () => {
     <section className="bg-white md:w-3/4 lg:w-3/5 mx-auto register-login animated fadeInDown faster">
       <div className="register content-center ">
         <div className="left">
-          <h1 className="lm-logo">
-            <span className="lm-lunch">LUNCH</span>
-            <span className="lm-memo">memo</span>
-          </h1>
+          <Logo />
           <p>
             Lunch Memo makes it easy to organize lunch with your colleagues and other professionals
             in your area. Spend less time finding lunch spots and scheduling with everyone and
