+++ pages/signup.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useRouter } from 'next/router';
 import Layout from '@components/layouts/Layout';
-import RegistrationPanel from '@components/layouts/RegistrationPanel';
+import RegistrationPanel from '@components/modules/registration/RegistrationPanel';
 import unAuthGuard from '@utils/unAuthGuard';
 
 const SignupPage: React.FC = () => {
